Fix minor issues in coverage tests

+++ coverage.sh
@@ -29,14 +29,15 @@ coverlogfile=$5
 # code coverage environment variables
 export GCOV_PREFIX=$coverdir
 export GCOV_PREFIX_STRIP=`echo $topdir | awk '{print gsub(/\//, "")}'`
+export COMPREPORT=YES
 
-LIBLIST=(iotp-device-coverage-as-lib skip iotp-gateway-coverage-as-lib iotp-application-coverage-as-lib iotp-managed-device-coverage-as-lib iotp-managed-gateway-coverage-as-lib)
+LIBLIST=(iotp-device-coverage-as-lib iotp-gateway-coverage-as-lib iotp-application-coverage-as-lib iotp-managed-device-coverage-as-lib iotp-managed-gateway-coverage-as-lib)
 
-# LIBLIST=(iotp-device-coverage-as-lib skip)
+# LIBLIST=(iotp-device-coverage-as-lib)
 
-TESTLIST=(config_tests_coverage device_tests_coverage gateway_tests_coverage application_tests_coverage managedDevice_tests_coverage managedGateway_tests_coverage)
+TESTLIST=(device_tests_coverage gateway_tests_coverage application_tests_coverage managedDevice_tests_coverage managedGateway_tests_coverage)
 
-# TESTLIST=(config_tests_coverage device_tests_coverage)
+# TESTLIST=(device_tests_coverage)
 
 if [ "${ostype}" == "Darwin" ]; then
     export DYLD_LIBRARY_PATH=$paholibdir:$coverdir
@@ -73,7 +74,6 @@ do
 
     echo
     echo "Coverage Test: ${tname}"
-    echo "Build test"
     make -C test $tname
 
     testname=`echo $tname | cut -d"_" -f1`
@@ -82,7 +82,7 @@ do
     cd ${coverdir}
 
     echo
-    echo "Run test"
+    echo "Run test ${tname}"
 
     if [ "${ostype}" == "Darwin" ]; then
         DYLD_LIBRARY_PATH=$paholibdir:$coverdir; ./$tname
@@ -91,7 +91,7 @@ do
     fi
 
     echo
-    echo "Coverage test report"
+    echo "Coverage report: ${tname}"
     mv build/coverage/*.gcda .
 
     tlog="${coverdir}/reports/${testname}.log"
@@ -108,8 +108,8 @@ do
     printf "%s\t\t%s\t%s\t%s\n" "FileName" "Lines" "LinesExecuted" "BranchesExecuted" >> $slog
     echo "-----------------------------------------------------------------" >> $slog
 
-    echo "gcov -a -b -c -u -gcno=${coverdir}/${testname}_tests.gcno -gcda=${coverdir}/${testname}_tests.gcda ${coverdir}/${testname}_tests.c" 
-    gcov -a -b -c -u -gcno=${coverdir}/${testname}_tests.gcno -gcda=${coverdir}/${testname}_tests.gcda ${coverdir}/${testname}_tests.c | tee -a ${tlog}
+    echo "gcov -a -b -c -u ${coverdir}/${testname}_tests.c" 
+    gcov -a -b -c -u ${coverdir}/${testname}_tests.c | tee -a ${tlog}
 
     # write summary
     nolines=`grep "Lines " ${tlog} | cut -d" " -f4`
@@ -129,31 +129,36 @@ do
     rm -f test_utils.gcno 
     rm -f ${testname}_tests.gcno
 
-    for gcno in `ls *.gcno`
-    do
-        fname=`echo $gcno | cut -d"." -f1`
-        echo "Process ${fname}"
-        echo "gcov -a -b -c -u -gcda=${coverdir}/${fname}.gcda -gcno=${coverdir}/${fname}.gcno src/wiotp/sdk/${fname}.c" 
-        gcov -a -b -c -u -gcda=${coverdir}/${fname}.gcda -gcno=${coverdir}/${fname}.gcno src/wiotp/sdk/${fname}.c | tee -a ${llog}
+    if [ "${COMPREPORT}" == "YES" ]; then
 
-        # write summary
-        nolines=`grep "Lines " ${llog} | cut -d" " -f4`
-        lExec=`grep "Lines executed" ${llog} | cut -d" " -f2 | cut -d":" -f2`
-        bExec=`grep "Branches executed" ${llog} | cut -d" " -f2 | cut -d":" -f2`
+        for gcno in `ls *.gcno`
+        do
+            fname=`echo $gcno | cut -d"." -f1`
+            echo "Process ${fname}"
+    
+            echo "gcov -a -b -c -u ${fname}.c" 
+            gcov -a -b -c -u ${fname}.c | tee -a ${llog}
 
-        if [ "${lExec}" != "0.00%" ] ; then
-            ccount=`echo "${fname}.c" | wc -c`
-            if [ $ccount -gt  16 ]; then 
-                printf "%s\t%s\t%s\t\t%s\n" "${fname}.c" "$nolines" "$lExec" "$bExec" >> $slog
-            else
-                printf "%s\t\t%s\t%s\t\t%s\n" "${fname}.c" "$nolines" "$lExec" "$bExec" >> $slog
+            # write summary
+            nolines=`grep "Lines " ${llog} | cut -d" " -f4`
+            lExec=`grep "Lines executed" ${llog} | cut -d" " -f2 | cut -d":" -f2`
+            bExec=`grep "Branches executed" ${llog} | cut -d" " -f2 | cut -d":" -f2`
+    
+            if [ "${lExec}" != "0.00%" ] ; then
+                ccount=`echo "${fname}.c" | wc -c`
+                if [ $ccount -gt  16 ]; then 
+                    printf "%s\t%s\t%s\t\t%s\n" "${fname}.c" "$nolines" "$lExec" "$bExec" >> $slog
+                else
+                    printf "%s\t\t%s\t%s\t\t%s\n" "${fname}.c" "$nolines" "$lExec" "$bExec" >> $slog
+                fi
             fi
-        fi
+    
+            rm -f ${llog}
+            touch ${llog}
+    
+        done
 
-        rm -f ${llog}
-        touch ${llog}
-
-    done
+    fi
 
     count=$(( $count + 1 ))
